+++ packages/markup/src/index.ts
@@ -18,6 +18,7 @@ export {
   stringifyBlock,
 } from "./helpers.js";
 export { markdown, markdownCodec } from "./markdown/index.js";
+export { remarkWikiLink } from "./markdown/wikilink.js";
 export { mdx, mdxCodec } from "./mdx/index.js";
 export type {
   AssetPathResolver,